+++ internal/docs_test.go
@@ -70,14 +70,33 @@ func (d DocDefinitions) RenderMarkdown(t *testing.T, repositoryPath string) []by
 		write(t, out, "### %s\n\n", el.Path)
 		write(t, out, "Type: `%s` <sup>[\\[ref\\]](%s/%s#L%d)</sup>\n\n", el.Type, repositoryPath, el.File, el.Line)
 
-		if d := el.Deprecated; len(d) > 0 {
-			write(t, out, "> [!WARNING]\n")
-			write(t, out, "> ***DEPRECATED***\n")
-			write(t, out, "> \n")
-			for _, line := range d {
-				write(t, out, "> **%s**\n", line)
+		if grade := el.Grade; grade != nil {
+			switch grade.Grade {
+			case DocDefinitionGradeDeprecated:
+				write(t, out, "> [!WARNING]\n")
+				write(t, out, "> ***DEPRECATED***\n")
+				write(t, out, "> \n")
+				for _, line := range grade.Message {
+					write(t, out, "> **%s**\n", line)
+				}
+				write(t, out, "\n")
+			case DocDefinitionGradeAlpha:
+				write(t, out, "> [!WARNING]\n")
+				write(t, out, "> ***ALPHA***\n")
+				write(t, out, "> \n")
+				for _, line := range grade.Message {
+					write(t, out, "> **%s**\n", line)
+				}
+				write(t, out, "\n")
+			case DocDefinitionGradeBeta:
+				write(t, out, "> [!IMPORTANT]\n")
+				write(t, out, "> ***BETA***\n")
+				write(t, out, "> \n")
+				for _, line := range grade.Message {
+					write(t, out, "> **%s**\n", line)
+				}
+				write(t, out, "\n")
 			}
-			write(t, out, "\n")
 		}
 
 		if d := el.Important; d != nil {