+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - (Bugfix) Fix Checksum Calculation
 - (Bugfix) Implement Missing Group Tolerations
 - (Feature) (Platform) OpenID Logout Endpoints
+- (Maintenance) Grade Doc Field
 
 ## [1.2.48](https://github.com/arangodb/kube-arangodb/tree/1.2.48) (2025-05-08)
 - (Maintenance) Extend Documentation